requirements.txt
fastapi>=0.109.0
python-multipart>=0.0.7
uvicorn>=0.27.0
jinja2>=3.1.3
nltk>=3.8.1
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0